# Conflicts:
#	bin/org/team708/robot/Constants.class
#	bin/org/team708/robot/Robot.class
#	bin/org/team708/robot/RobotMap.class
#
bin/org/team708/robot/commands/drivetrain/DriveStraightToEncoderDistance
.class
#
bin/org/team708/robot/commands/drivetrain/DriveStraightToEncoderDistance
OrTime.class
#	bin/org/team708/robot/commands/drivetrain/DriveWithOpticalSensor.class
#	bin/org/team708/robot/subsystems/Drivetrain.class
#	src/org/team708/robot/Robot.java

+++ src/org/team708/robot/util/Potentiometer.java
@@ -18,9 +18,9 @@ public class Potentiometer {
     private final int MIN_POTENTIOMETER_VOLTAGE = 0;
     private final int MAX_POTENTIOMETER_VOLTAGE = 5;
     private final int MIN_POTENTIOMETER_ANGLE = 0;
-    private 	  int maxPotentiometerAngle;
+    private 	  double maxPotentiometerAngle;
     
-    public Potentiometer(int pwmPort, int numberOfRotations) {
+    public Potentiometer(int pwmPort, double numberOfRotations) {
         potentiometer = new AnalogInput(pwmPort);
         maxPotentiometerAngle = (numberOfRotations * 360);
     }
